Implement a list method
This change implements a barebones list method to let users see what files exist on their
drive (a nice complement to the pull function).

Right now it just dumps a list of files and directories:
If the path is a file then it is listed.
If the path is a directory then the direct children are listed.